move action to content
