+++ apps/localplanning.services/src/components/OptionCard.astro
@@ -1,7 +1,7 @@
 ---
 import Button from "@components/Button.astro";
 import { actions } from "@content/actions";
-import type { Action } from "@stores/action";
+import type { Action } from "@content/action";
 
 interface Props {
   action: Action;
