+++ apps/localplanning.services/src/layouts/SearchLayout.astro
@@ -4,7 +4,7 @@ import Container from "@components/Container.astro";
 import Masthead from "@components/Masthead.astro";
 import AddressSearch from "@components/AddressSearch.tsx";
 import ApplicationsBanner from "@components/ApplicationsBanner.astro";
-import type { Action } from "@stores/action";
+import type { Action } from "@content/action";
 
 interface Props {
   title: string;
