+++ apps/localplanning.services/src/pages/index.astro
@@ -4,7 +4,7 @@ import Container from "@components/Container.astro";
 import OptionCard from "@components/OptionCard.astro";
 import Masthead from "@components/Masthead.astro";
 import ApplicationsBanner from "@components/ApplicationsBanner.astro";
-import { ACTIONS } from "@stores/action";
+import { ACTIONS } from "@content/action";
 import { Image } from "astro:assets";
 import lpaMap from "/public/planx-lpa-map.svg";
